src/services/mcp/requirements.txt
tensorflow==2.15.0
tensorflow-serving-api==2.15.0
grpcio==1.59.3
protobuf==4.25.1
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0 